msm: kgsl: Remove "runtime" per-process pagetable checks
There was a time when per-process pagetable support depended on a
number of HW and SW factors but those days are gone. All supported
targets support per-process pagetables. The only remaining control
is the Kconfig flag. Remove much of the infrastructure used to
determine per-process pagetable support and just use the static

By the same token, kgsl_mmu_use_cpu_map() also boils down to
depending on the per-process Kconfig so remove all that
infrastructure and return kgsl_mmu_is_perprocess() for that
function.

I realize I could take it another step and remove
kgsl_mmu_is_perprocess() and kgsl_mmu_use_cpu_map() all together
but that would be much harder to cherry-pick if we need to.
All the functions are inline the compiler should be able to do
the right thing at the expense of slightly less pretty code.

Change-Id: Ic0dedbade36c59376e78d7c2c31b7691482096f6
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>

Conflicts:
	drivers/gpu/msm/kgsl_iommu.c
	drivers/gpu/msm/kgsl_iommu.h
	drivers/gpu/msm/kgsl_mmu.h