gpu: ion: Optimize the vmap step size in ion page zeroing
When ion zeroes out pages it uses a vmap step size equal to
the number of pages we map into vmalloc space once in order
to memset 0 the allocation. Till now, the step size was 1/8th
the vmalloc space. This has one side-effect - other applications
in the system might starve due to lack of vmalloc space.
We conducted experiments with the goal to reduce the step
size while not hurting performance, and arrived at step size
of 2MB.

based on: 4e121e137f4277a4794ce16580b03270f22b0d9f

Signed-off-by: Kostas K <mpokwsths@gmail.com>